test: move initialize_graph test
