+++ tests/unit/conftest.py
@@ -5,6 +5,7 @@ import pytest
 
 from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
 from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
+from expertsystem.topology import Topology
 from expertsystem.ui import (
     InteractionTypes,
     StateTransitionManager,
@@ -59,3 +60,26 @@ def jpsi_to_gamma_pi_pi_helicity_amplitude_generator(
     amplitude_generator = HelicityAmplitudeGenerator()
     amplitude_generator.generate(jpsi_to_gamma_pi_pi_helicity_solutions)
     return amplitude_generator
+
+
+@pytest.fixture(scope="package")
+def dummy_topology():
+    r"""Create a dummy `Topology`.
+
+    Has the following shape:
+
+    .. code-block::
+
+        e0 -- (N0) -- e1 -- (N1) -- e3
+                \             \
+                 e2            e4
+    """
+    topology = Topology()
+    topology.add_node(0)
+    topology.add_node(1)
+    topology.add_edges([0, 1, 2, 3, 4])
+    topology.attach_edges_to_node_ingoing([0], 0)
+    topology.attach_edges_to_node_ingoing([1], 1)
+    topology.attach_edges_to_node_outgoing([1, 2], 0)
+    topology.attach_edges_to_node_outgoing([3, 4], 1)
+    return topology
